閲覧権限の機能
