+++ front/src/hooks/role/index.ts
@@ -1,0 +1,1 @@
+export { useRoleEdit } from '@/hooks/role/useRoleEdit'
